Implement "Shield Barrier" trait
A new trait "Shield Barrier" has been added for WAR and DRK!
Reduces damage while holding a shield.
Level 30/45/60/75.
Reduction 10/15/20/25

+++ src/map/lua/lua_baseentity.cpp
@@ -14265,7 +14265,7 @@ inline int32 CLuaBaseEntity::hasTrait(lua_State *L)
         return 1;
     }
 
-    lua_pushboolean(L, charutils::hasTrait((CCharEntity*)m_PBaseEntity, (uint8)lua_tointeger(L, 1)));
+    lua_pushboolean(L, charutils::hasTrait((CCharEntity*)m_PBaseEntity, (uint16)lua_tointeger(L, 1)));
     return 1;
 }
